asociacion cliente y usuarios rebase
