+++ casa_de_cambios/urls.py
@@ -1,6 +1,6 @@
+# global-exchange/urls.py
 from django.contrib import admin
 from django.urls import path, include
-from users import views as user_views
 from django.shortcuts import redirect
 
 urlpatterns = [
@@ -8,4 +8,6 @@ urlpatterns = [
     path("", include("interfaz.urls")),
     path('usuarios/', include('users.urls')),  # URLs de la app users
     path('', lambda request: redirect('seleccionar_cliente')),  # raíz redirige a seleccionar cliente
+    path('nueva_ruta/', include('nueva_app.urls')),  # Incluye las URLs de la nueva app
+    path('', lambda request: redirect('nueva_ruta/')),  # Redirige a la nueva app
 ]
