+++ nueva_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class NuevaAppConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'nueva_app'
