+++ nueva_app/templates/admin_asociar.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Asociación de Clientes</title>
+</head>
+<body>
+    <h1>Asociar Clientes a Usuarios</h1>
+
+    <form method="post" action="{% url 'asociar_admin' %}">
+        {% csrf_token %}
+        <label for="usuario">Seleccionar Usuario:</label>
+        <select name="usuario" id="usuario">
+            {% for usuario in usuarios %}
+                <option value="{{ usuario.id }}">{{ usuario.username }}</option>
+            {% endfor %}
+        </select>
+        <br><br>
+
+        <label>Seleccionar Clientes:</label>
+        <br>
+        {% for cliente in clientes %}
+            <div>
+                <input type="checkbox" name="clientes" value="{{ cliente.id }}">
+                <label>{{ cliente.nombre_completo }} (Cédula: {{ cliente.cedula }})</label>
+            </div>
+        {% endfor %}
+        <br>
+        <button type="submit">Asociar Clientes</button>
+    </form>
+
+</body>
+</html>
