+++ nueva_app/templates/error.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Error</title>
+</head>
+<body>
+    <h1>Error</h1>
+    <p>{{ mensaje }}</p>
+</body>
+</html>
