--- users/views.py
@@ -1,90 +1,0 @@
-# users/views.py
-from django.shortcuts import render, redirect
-from django.contrib.auth.decorators import login_required
-from django.contrib.auth.models import User
-from .models import Cliente, AsignacionCliente
-
-
-@login_required
-def seleccion_cliente_view(request):
-    """
-    Vista para que el usuario seleccione uno de sus clientes asignados.
-    """
-    clientes = Cliente.objects.filter(usuarios__id=request.user.id)
-
-    if not clientes.exists():
-        # Si el usuario no tiene clientes asignados
-        return render(request, 'error.html',
-                      {'mensaje': 'No tienes clientes asignados. Por favor, contacta a un administrador.'})
-
-    if clientes.count() == 1:
-        # Si solo tiene un cliente, se selecciona automáticamente
-        request.session['cliente_id_actual'] = clientes.first().id
-        return redirect('home')
-
-    # Si tiene varios clientes, se muestra la lista para elegir
-    return render(request, 'asociar_clientes.html', {'clientes': clientes})
-
-
-@login_required
-def asociar_admin_view(request):
-    """
-    Vista para que el administrador asigne clientes a usuarios.
-    """
-    if not request.user.is_staff:
-        return redirect('home')
-
-    if request.method == 'POST':
-        # El admin envía la asociación de clientes a un usuario
-        usuario_id = request.POST.get('usuario')
-        clientes_ids = request.POST.getlist('clientes')
-
-        try:
-            usuario = User.objects.get(id=usuario_id)
-            # Elimina asociaciones anteriores
-            AsignacionCliente.objects.filter(usuario=usuario).delete()
-
-            # Crea las nuevas asociaciones
-            for cliente_id in clientes_ids:
-                cliente = Cliente.objects.get(id=cliente_id)
-                AsignacionCliente.objects.create(usuario=usuario, cliente=cliente)
-
-            return redirect('asociar_admin')
-
-        except (User.DoesNotExist, Cliente.DoesNotExist):
-            return render(request, 'error.html', {'mensaje': 'Error: Usuario o cliente no encontrado.'})
-
-    else:
-        # Muestra formulario de asignación
-        usuarios = User.objects.all()
-        clientes = Cliente.objects.all()
-        return render(request, 'admin_asociar.html', {'usuarios': usuarios, 'clientes': clientes})
-
-
-def home_view(request):
-    """
-    Vista principal: muestra el cliente actual o redirige a la selección.
-    """
-    if request.user.is_authenticated:
-        if 'cliente_id_actual' not in request.session:
-            return redirect('seleccionar_cliente')
-
-        cliente_actual = Cliente.objects.get(id=request.session['cliente_id_actual'])
-        return render(request, 'home.html', {'cliente_actual': cliente_actual})
-
-    return render(request, 'landing_page.html')
-
-
-def guardar_seleccion_cliente(request, cliente_id):
-    """
-    Guarda en la sesión el cliente seleccionado por el usuario.
-    """
-    if request.user.is_authenticated:
-        try:
-            cliente = Cliente.objects.get(id=cliente_id, usuarios__id=request.user.id)
-            request.session['cliente_id_actual'] = cliente.id
-            return redirect('home')
-        except Cliente.DoesNotExist:
-            return render(request, 'error.html', {'mensaje': 'El cliente seleccionado no es válido.'})
-
-    return redirect('login')
